requirements.txt
PyPDF2
langchain
torch
numpy
huggingface-hub
huggingface
python-dotenv
faiss-cpu
InstructorEmbedding
transformers
sentence_transformers